refactor: separate GetGenresUseCase.kt into two different usecases for movie and tv show
