+++ domain/src/main/java/com/baghdad/domain/usecase/movie/GetMovieGenresUseCase.kt
@@ -1,0 +1,13 @@
+package com.baghdad.domain.usecase.movie
+
+import com.baghdad.domain.repository.MovieRepository
+import com.baghdad.entity.media.Genre
+import javax.inject.Inject
+
+class GetMovieGenresUseCase @Inject constructor(
+    private val movieRepository: MovieRepository,
+) {
+    suspend fun getMovieGenres(): List<Genre> {
+        return movieRepository.getGenres()
+    }
+}
