+++ domain/src/main/java/com/baghdad/domain/usecase/tvShow/GetTvShowGenresUseCase.kt
@@ -1,0 +1,13 @@
+package com.baghdad.domain.usecase.tvShow
+
+import com.baghdad.domain.repository.TvShowRepository
+import com.baghdad.entity.media.Genre
+import javax.inject.Inject
+
+class GetTvShowGenresUseCase @Inject constructor(
+    private val tvShowRepository: TvShowRepository,
+) {
+    suspend fun getTvShowGenres(): List<Genre> {
+        return tvShowRepository.getGenres()
+    }
+}
