+++ viewmodel/src/main/java/com/baghdad/viewmodel/trendingMovie/TrendingMoviesViewModel.kt
@@ -2,8 +2,8 @@ package com.baghdad.viewmodel.trendingMovie
 
 import androidx.paging.PagingData
 import com.baghdad.domain.exception.NoInternetException
-import com.baghdad.domain.usecase.genre.GetGenresUseCase
 import com.baghdad.domain.usecase.login.IsUserLoggedInUseCase
+import com.baghdad.domain.usecase.movie.GetMovieGenresUseCase
 import com.baghdad.domain.usecase.movie.GetTrendingMoviesUseCase
 import com.baghdad.domain.usecase.savedList.AddMovieToSavedListUseCase
 import com.baghdad.domain.usecase.savedList.CreateSavedListUseCase
@@ -25,7 +25,7 @@ import javax.inject.Inject
 @HiltViewModel
 class TrendingMoviesViewModel @Inject constructor(
     private val getTrendingMoviesUseCase: GetTrendingMoviesUseCase,
-    private val getGenresUseCase: GetGenresUseCase,
+    private val getMovieGenresUseCase: GetMovieGenresUseCase,
     private val isUserLoggedInUseCase: IsUserLoggedInUseCase,
     private val getSavedListsUseCase: GetSavedListsUseCase,
     private val addMovieToSavedListUseCase: AddMovieToSavedListUseCase,
@@ -85,7 +85,7 @@ class TrendingMoviesViewModel @Inject constructor(
 
     private fun loadGenres() {
         tryToExecute(
-            callee = { getGenresUseCase.getMovieGenres() },
+            callee = { getMovieGenresUseCase.getMovieGenres() },
             onSuccess = ::handleGenreSuccess,
             onError = ::onLoadDataError,
             dispatcher = defaultDispatcher
